PM-24609 Removed deprecated function in CipherMatchingHelper. Moved Singlular match tests to the original CipherMatchingHelperTests. And fixed some warnings.
