+++ BitwardenKit/UI/Platform/Application/Appearance/UI.swift
@@ -11,7 +11,7 @@ public enum UI {
     // MARK: Utilities
 
     /// App-wide flag that allows disabling UI animations for testing.
-    nonisolated(unsafe) public static var animated = true
+    public nonisolated(unsafe) static var animated = true
 
     /// The language code at initialization.
     public static var initialLanguageCode: String? {
@@ -25,7 +25,7 @@ public enum UI {
 
     #if DEBUG
     /// App-wide flag that allows overriding the OS level sizeCategory for testing.
-    nonisolated(unsafe) public static var sizeCategory: UIContentSizeCategory?
+    public nonisolated(unsafe) static var sizeCategory: UIContentSizeCategory?
     #endif
 
     // MARK: Factories
