+++ BitwardenShared/UI/Platform/Settings/Settings/Other/OtherSettingsViewTests+SnapshotTests.swift
@@ -1,4 +1,3 @@
-// swiftlint:disable:this file_name
 import BitwardenKit
 import BitwardenKitMocks
 import BitwardenResources
